Changed i64 and u64 to 128 bit respectively

+++ src/impls/int.rs
@@ -1,6 +1,6 @@
 use crate::PmlElem;
 
-impl<'a> From<&'a PmlElem> for &'a i64 {
+impl<'a> From<&'a PmlElem> for &'a i128 {
     fn from(elem: &'a PmlElem) -> Self {
         match elem {
             PmlElem::PmlInt(i) => i,
@@ -9,7 +9,7 @@ impl<'a> From<&'a PmlElem> for &'a i64 {
     }
 }
 
-impl Into<PmlElem> for i64 {
+impl Into<PmlElem> for i128 {
     fn into(self) -> PmlElem {
         PmlElem::PmlInt(self)
     }
